functions/ai/requirements.txt
functions-framework==3.5.0
google-cloud-secret-manager==2.18.1
google-cloud-storage==2.10.0
google-cloud-translate==3.12.1
google-cloud-speech==2.21.0
google-generativeai==0.3.2
PyJWT==2.8.0
flask==3.0.0
flask-cors==4.0.0
cachetools==5.3.2
requests==2.31.0
python-multipart==0.0.6
Pillow==10.1.0
pydantic==2.5.2